extract interface for compile job
