+++ server/src/test/DbtCompileJob.spec.ts
@@ -1,8 +1,8 @@
 import { assertThat } from 'hamjest';
 import { err, ok } from 'neverthrow';
 import { instance, mock, verify, when } from 'ts-mockito';
-import { DbtCompileJob } from '../DbtCompileJob';
 import { CompileResponse, DbtRpcClient, PollResponse } from '../DbtRpcClient';
+import { DbtRpcCompileJob } from '../DbtRpcCompileJob';
 
 describe('DbtCompileJob', () => {
   const MODEL = 'test_model';
@@ -71,12 +71,12 @@ describe('DbtCompileJob', () => {
   }
 
   beforeEach(() => {
-    DbtCompileJob.COMPILE_MODEL_MAX_RETRIES = 10;
-    DbtCompileJob.COMPILE_MODEL_TIMEOUT_MS = 0;
+    DbtRpcCompileJob.COMPILE_MODEL_MAX_RETRIES = 10;
+    DbtRpcCompileJob.COMPILE_MODEL_TIMEOUT_MS = 0;
 
-    DbtCompileJob.POLL_MAX_RETRIES = 86;
-    DbtCompileJob.POLL_TIMEOUT_MS = 0;
-    DbtCompileJob.MAX_RETRIES_FOR_UNKNOWN_ERROR = 5;
+    DbtRpcCompileJob.POLL_MAX_RETRIES = 86;
+    DbtRpcCompileJob.POLL_TIMEOUT_MS = 0;
+    DbtRpcCompileJob.MAX_RETRIES_FOR_UNKNOWN_ERROR = 5;
   });
 
   it('Should return ok result', async () => {
@@ -86,7 +86,7 @@ describe('DbtCompileJob', () => {
     when(mockDbtRpcClient.compile(MODEL)).thenReturn(compileModelSuccess());
     when(mockDbtRpcClient.pollOnceCompileResult(TOKEN)).thenReturn(pollOnceCompileResultSuccess());
 
-    const compileJob = new DbtCompileJob(instance(mockDbtRpcClient), MODEL);
+    const compileJob = new DbtRpcCompileJob(instance(mockDbtRpcClient), MODEL);
 
     // act
     await compileJob.start();
@@ -103,7 +103,7 @@ describe('DbtCompileJob', () => {
     when(mockDbtRpcClient.compile(MODEL)).thenReturn(compileModelError()).thenReturn(compileModelSuccess());
     when(mockDbtRpcClient.pollOnceCompileResult(TOKEN)).thenReturn(pollOnceCompileResultSuccess());
 
-    const compileJob = new DbtCompileJob(instance(mockDbtRpcClient), MODEL);
+    const compileJob = new DbtRpcCompileJob(instance(mockDbtRpcClient), MODEL);
 
     // act
     await compileJob.start();
@@ -121,7 +121,7 @@ describe('DbtCompileJob', () => {
     when(mockDbtRpcClient.compile(MODEL)).thenReturn(Promise.resolve(undefined)).thenReturn(compileModelSuccess());
     when(mockDbtRpcClient.pollOnceCompileResult(TOKEN)).thenReturn(pollOnceCompileResultSuccess());
 
-    const compileJob = new DbtCompileJob(instance(mockDbtRpcClient), MODEL);
+    const compileJob = new DbtRpcCompileJob(instance(mockDbtRpcClient), MODEL);
 
     // act
     await compileJob.start();
@@ -134,53 +134,53 @@ describe('DbtCompileJob', () => {
   it('Should retry compileModel max times and fail due to network error', async () => {
     // arrange
     const mockDbtRpcClient = mock(DbtRpcClient);
-    DbtCompileJob.COMPILE_MODEL_MAX_RETRIES = 1;
+    DbtRpcCompileJob.COMPILE_MODEL_MAX_RETRIES = 1;
 
     when(mockDbtRpcClient.compile(MODEL)).thenReturn(Promise.resolve(undefined));
 
-    const compileJob = new DbtCompileJob(instance(mockDbtRpcClient), MODEL);
+    const compileJob = new DbtRpcCompileJob(instance(mockDbtRpcClient), MODEL);
 
     // act
     await compileJob.start();
 
     // assert
-    assertThat(compileJob.result, err(DbtCompileJob.NETWORK_ERROR));
-    verify(mockDbtRpcClient.compile(MODEL)).times(DbtCompileJob.COMPILE_MODEL_MAX_RETRIES + 1);
+    assertThat(compileJob.result, err(DbtRpcCompileJob.NETWORK_ERROR));
+    verify(mockDbtRpcClient.compile(MODEL)).times(DbtRpcCompileJob.COMPILE_MODEL_MAX_RETRIES + 1);
   });
 
   it('Should retry compileModel max times and fail due to dbt-rpc error', async () => {
     // arrange
     const mockDbtRpcClient = mock(DbtRpcClient);
-    DbtCompileJob.COMPILE_MODEL_MAX_RETRIES = 1;
+    DbtRpcCompileJob.COMPILE_MODEL_MAX_RETRIES = 1;
 
     when(mockDbtRpcClient.compile(MODEL)).thenReturn(Promise.resolve(undefined)).thenReturn(compileModelError());
 
-    const compileJob = new DbtCompileJob(instance(mockDbtRpcClient), MODEL);
+    const compileJob = new DbtRpcCompileJob(instance(mockDbtRpcClient), MODEL);
 
     // act
     await compileJob.start();
 
     // assert
     assertThat(compileJob.result, err(COMPILE_MODEL_ERROR));
-    verify(mockDbtRpcClient.compile(MODEL)).times(DbtCompileJob.COMPILE_MODEL_MAX_RETRIES + 1);
+    verify(mockDbtRpcClient.compile(MODEL)).times(DbtRpcCompileJob.COMPILE_MODEL_MAX_RETRIES + 1);
   });
 
   it('Should stop compileModel during retrying', async () => {
     // arrange
     const mockDbtRpcClient = mock(DbtRpcClient);
-    DbtCompileJob.COMPILE_MODEL_TIMEOUT_MS = 300;
+    DbtRpcCompileJob.COMPILE_MODEL_TIMEOUT_MS = 300;
 
     when(mockDbtRpcClient.compile(MODEL)).thenReturn(Promise.resolve(compileModelSuccess())).thenReturn(compileModelSuccess());
 
-    const compileJob = new DbtCompileJob(instance(mockDbtRpcClient), MODEL);
+    const compileJob = new DbtRpcCompileJob(instance(mockDbtRpcClient), MODEL);
     const startPromise = compileJob.start();
 
     // act
-    await compileJob.stop();
+    await compileJob.forceStop();
     await startPromise;
 
     // assert
-    assertThat(compileJob.result, err(DbtCompileJob.STOP_ERROR));
+    assertThat(compileJob.result, err(DbtRpcCompileJob.STOP_ERROR));
     verify(mockDbtRpcClient.compile(MODEL)).once();
   });
 
@@ -191,7 +191,7 @@ describe('DbtCompileJob', () => {
     when(mockDbtRpcClient.compile(MODEL)).thenReturn(compileModelSuccess());
     when(mockDbtRpcClient.pollOnceCompileResult(TOKEN)).thenReturn(pollOnceCompileResultRunning()).thenReturn(pollOnceCompileResultSuccess());
 
-    const compileJob = new DbtCompileJob(instance(mockDbtRpcClient), MODEL);
+    const compileJob = new DbtRpcCompileJob(instance(mockDbtRpcClient), MODEL);
 
     // act
     await compileJob.start();
@@ -209,7 +209,7 @@ describe('DbtCompileJob', () => {
     when(mockDbtRpcClient.compile(MODEL)).thenReturn(compileModelSuccess());
     when(mockDbtRpcClient.pollOnceCompileResult(TOKEN)).thenReturn(Promise.resolve(undefined)).thenReturn(pollOnceCompileResultSuccess());
 
-    const compileJob = new DbtCompileJob(instance(mockDbtRpcClient), MODEL);
+    const compileJob = new DbtRpcCompileJob(instance(mockDbtRpcClient), MODEL);
 
     // act
     await compileJob.start();
@@ -223,20 +223,20 @@ describe('DbtCompileJob', () => {
   it('Should retry maximum MAX_RETRIES_FOR_UNKNOWN_ERROR in case of pollOnceCompileResultRunning returns undefined', async () => {
     // arrange
     const mockDbtRpcClient = mock(DbtRpcClient);
-    DbtCompileJob.MAX_RETRIES_FOR_UNKNOWN_ERROR = 2;
+    DbtRpcCompileJob.MAX_RETRIES_FOR_UNKNOWN_ERROR = 2;
 
     when(mockDbtRpcClient.compile(MODEL)).thenReturn(compileModelSuccess());
     when(mockDbtRpcClient.pollOnceCompileResult(TOKEN)).thenReturn(Promise.resolve(undefined));
 
-    const compileJob = new DbtCompileJob(instance(mockDbtRpcClient), MODEL);
+    const compileJob = new DbtRpcCompileJob(instance(mockDbtRpcClient), MODEL);
 
     // act
     await compileJob.start();
 
     // assert
-    assertThat(compileJob.result, err(DbtCompileJob.NETWORK_ERROR));
+    assertThat(compileJob.result, err(DbtRpcCompileJob.NETWORK_ERROR));
     verify(mockDbtRpcClient.compile(MODEL)).times(1);
-    verify(mockDbtRpcClient.pollOnceCompileResult(TOKEN)).times(DbtCompileJob.MAX_RETRIES_FOR_UNKNOWN_ERROR + 1);
+    verify(mockDbtRpcClient.pollOnceCompileResult(TOKEN)).times(DbtRpcCompileJob.MAX_RETRIES_FOR_UNKNOWN_ERROR + 1);
   });
 
   it('Should return error in case of pollOnceCompileResultRunning returns error', async () => {
@@ -246,7 +246,7 @@ describe('DbtCompileJob', () => {
     when(mockDbtRpcClient.compile(MODEL)).thenReturn(compileModelSuccess());
     when(mockDbtRpcClient.pollOnceCompileResult(TOKEN)).thenReturn(pollOnceCompileResultError());
 
-    const compileJob = new DbtCompileJob(instance(mockDbtRpcClient), MODEL);
+    const compileJob = new DbtRpcCompileJob(instance(mockDbtRpcClient), MODEL);
 
     // act
     await compileJob.start();
